layer clean up
