+++ src_vue/config/config.json
@@ -1,0 +1,28 @@
+{
+	"flask_host" : "0.0.0.0",
+
+	"collections_info_backend" : {
+		"default" : "mysql://root:root@127.0.0.1:3377/scot",
+		"en_books" : "mysql://root:root@127.0.0.1:3377/scot",
+		"fi_news" : "mysql://root:root@127.0.0.1:3306/DT_finnews_dependency",
+		"de70" : "mysql://root:root@127.0.0.1:3306/DT_de70mlcc_dependency"
+		
+
+	},
+
+	"collections_info_elastic" : {
+		"fi_news" : {"es_index": "fi_news", "es_host": "127.0.0.1", "es_port": 9292},
+		"de70" : {"es_index": "de70", "es_host": "127.0.0.1", "es_port": 9292}
+	},
+
+	"collections_info_frontend":{
+		
+		"English Books": {"key":"en_books", "target": "crisis/NN", "p": 100, "d": 20, "info": "English Google Books, 1520-2008."},
+			"German News 70m Dep": {"key": "de70", "target": "Krise#NN", "p": 100, "d": 8, "info": "German Online News - 70 Million Random Sentences, 1995-2010 "},
+			"Finnish News Dep": {"key": "fi_news", "target": "Yksi#NUM", "p": 100, "d": 8, "info": "Various Finnish Online News - 14 Million Sentences, "}
+		
+
+	}
+	
+	
+}
