+++ src_vue/services/chineseWhispers.py
